Создает новый презентер и переносит в него логику отрисовки отдельной карточки фильма
